Create README.mdv
Welcome to the Queen-Victorian project. This project was created to consolidate my first steps in the world of programming and to apply my knowledge. In this repository you will find basic web programming projects that include HTML, CSS, and JavaScript files.
## Goals
- Create web pages with simple functionality.
- Add visual and colorful design using CSS animations.
- Deepen JavaScript using async/await.
## Plans
In the future, I would like to add:
- Calculation tool.
- Simple game via JavaScript.
- A web page using React and Node.js.
-In the future, I plan to develop a Saas (Software as a Service) project that aims to solve real-world problems and help people. This idea will serve as the foundation for a business where I can apply my programming skills and create valuable services.

The goal is not only to enhance technical skills but also to support and empower individuals in their growth and development. My Saas project will become a reliable tool for those seeking assistance and expertise.

This is my learning journey, and I share it to inspire other beginner programmers as well.

## Thank you
Thank you for visiting my vault. If you have any ideas or suggestions, I would love to hear from you.